refactor(core): extract bin break grids and split stats R-parity suites
Move ggplot2 break-grid helpers (binBreaksWidth/Bins, binIndexOf) into
stats/bin-breaks.ts so statBin owns counting and emission only. Co-locate
the 376-line stats-parity monolith under tests/stats-parity/{qt,bin,
smooth,boxplot,summary,density} with shared fixtures (decision 0010 path
updated).
