+++ .changeset/core-stats-bin-breaks-parity.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+refactor: extract bin break grids and co-locate stats R-parity suites
