+++ docs/decisions/0010-m2-stats-notes.md
@@ -15,7 +15,7 @@ All fixtures are generated by
 `packages/core/tests/fixtures/stats/generate.R` (ggplot2 4.0.3, R 4.6.1)
 through the public `layer_data()` API (bin/smooth/boxplot/summary) or
 `stats::density()` directly, and asserted in
-`packages/core/tests/stats-parity.test.ts`. Measured maxima, then the
+`packages/core/tests/stats-parity/`. Measured maxima, then the
 asserted bound:
 
 | Stat                                                                                                             | Reference                        | Measured max deviation                                              | Asserted       |
